Fix MSVC build

+++ libarchive/archive_read_support_format_zip.c
@@ -937,7 +937,7 @@ zip_read_local_file_header(struct archive_read *a, struct archive_entry *entry,
 	archive_entry_set_atime(entry, zip_entry->atime, 0);
 
 	if ((zip->entry->mode & AE_IFMT) == AE_IFLNK) {
-		size_t linkname_length = zip_entry->compressed_size;
+		size_t linkname_length = (size_t)zip_entry->compressed_size;
 
 		archive_entry_set_size(entry, 0);
 		p = __archive_read_ahead(a, linkname_length, NULL);
@@ -2348,7 +2348,7 @@ read_zip64_eocd(struct archive_read *a, struct zip *zip, const char *p)
 	eocd64_size = archive_le64dec(p + 4) + 12;
 	if (eocd64_size < 56 || eocd64_size > 16384)
 		return;
-	if ((p = __archive_read_ahead(a, eocd64_size, NULL)) == NULL)
+	if ((p = __archive_read_ahead(a, (size_t)eocd64_size, NULL)) == NULL)
 		return;
 
 	/* Sanity-check the EOCD64 */
